+++ src/zio.zig
@@ -39,6 +39,7 @@ pub const Cancelable = common.Cancelable;
 pub const Timeoutable = common.Timeoutable;
 pub const Closeable = common.Closeable;
 pub const blockInPlace = common.blockInPlace;
+pub const blockInPlaceReserved = common.blockInPlaceReserved;
 
 pub const time = @import("time.zig"); // TODO: make non-pub
 pub const Duration = time.Duration;